BZ cleaning up files

+++ algorithms/util.cpp
@@ -5,8 +5,8 @@
 void printIntArray(std::vector<int> myArr)
 {
   int x;
-  
-  for(x = 0; x < myArr.size(); x++)
+
+  for(x = 0; x < (int)myArr.size(); x++)
   {
     std::cout << myArr[x] << " ";
   }
